Split config/panel.py over multiple files
